+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/goyek/goyek/compare/v0.6.0...HEAD)
+## [Unreleased](https://github.com/goyek/goyek/compare/v0.6.1...HEAD)
+
+## [0.6.1](https://github.com/goyek/goyek/compare/v0.6.0...v0.6.1) - 2021-12-27
+
+This release adds the possibility to change the defaults of the global parameters.
 
 ### Added